* coders/jp2.c: Change how Jasper is intialized.  Hopefully eliminate memory leak when reading ICC color profile. * Overall: Compilation warning elimination with GCC 4.0.1. * magick/utility.c (SystemCommand): Always report error status if a command fails.

+++ coders/meta.c
@@ -1147,8 +1147,9 @@ static Image *ReadMETAImage(const ImageInfo *image_info,
             }
           AttachBlob(iptc->blob,pinfo->info,pinfo->length);
           (void) LogMagickEvent(CoderEvent,GetMagickModule(),
-                                "META CODER APP1JPEG embed: 0x%08lx (%d)",
-                                (unsigned long)pinfo->info, pinfo->length);
+                                "META CODER APP1JPEG embed: 0x%08lx (%ld)",
+                                (unsigned long) pinfo->info,
+                                (long) pinfo->length);
           result=jpeg_embed(image,buff,iptc);
           DetachBlob(iptc->blob);
           DestroyImage(iptc);
